migrate整理とseed途中

+++ db/migrate/20201222115329_create_couse_parameters.rb
@@ -3,8 +3,8 @@ class CreateCouseParameters < ActiveRecord::Migration[5.1]
     create_table :couse_parameters do |t|
       t.references :user, foreign_key: true
       t.references :couse, foreign_key: true
-      t.integer :speed, null: false, default: 1
-      t.integer :power, null: false, default: 1
+      t.float :speed, null: false, default: 1
+      t.float :power, null: false, default: 1
       t.timestamps
     end
   end
